New Development: no             JIRA Issue: no       Ready to Release: not until further interactive testing is done      Interface Changes: none What Interface Changed: none          Test Programs: none   Put in Release Notes: no
           Description: major simplification of region code, moving from a division of Qt aware VS GUI independent to full integration of Qt.



git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@22533 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/display/DisplayDatas/LatticePADD.tcc
@@ -1217,7 +1217,7 @@ void LatticePADisplayData<T>::drawBeamEllipse_(WorldCanvas* wc) {
 
     if( beams_.size( ) == 0 || beamOnOff_->value()=="No" ) return;
 
-    vector<double> beam = beams_.size( ) > activeZIndex_ ? beams_[activeZIndex_] : beams_[0];
+    vector<double> beam = (Int) beams_.size( ) > activeZIndex_ ? beams_[activeZIndex_] : beams_[0];
     // The major and minor axes must both be positive
     if ( beam[0] <= 0.0 || beam[1] <= 0.0 ) return;
